backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
PyJWT
python-dotenv
mysqlclient
python-docx>=0.8.11
docxtpl>=0.16.4
docxcompose>=1.4.0